Reconcile the definitions of Core.
