+++ src/Path/REPL.hs
@@ -19,6 +19,7 @@ import qualified Data.Set as Set
 import Data.Traversable (for)
 import Data.Void
 import GHC.Generics (Generic1)
+import Path.Core
 import Path.Elab
 import Path.Error
 import Path.Module as Module
